oss-knowledge-search/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database Connectors
qdrant-client==1.15.1
neo4j==5.15.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Azure OpenAI
openai==1.5.0
azure-identity==1.15.0

# Data Processing
numpy==1.24.3
pandas==2.0.3

# HTTP and Networking
httpx==0.25.2
requests==2.31.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0